+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Debug helper for sources with a invalid path
 
 ### Fixed
+
 - Issue when the theme value has no `src` key.
 
 ## [1.2.3] - 2025-11-20