+++ CHANGELOG.md
@@ -7,7 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[unreleased]: https://github.com/hyva-themes/hyva-modules-tailwind-js/compare/1.2.3...HEAD
+[unreleased]: https://github.com/hyva-themes/hyva-modules-tailwind-js/compare/1.2.4...HEAD
+
+## [1.2.4] - 2025-11-20
+
+[1.2.4]: https://github.com/hyva-themes/hyva-modules-tailwind-js/compare/1.2.3...1.2.4
+
+### Added
+
+- Debug helper for .gitignore issues with Tailwind v4 and sources
+- Debug helper for sources with a invalid path
+
+### Fixed
+
+- Issue when the theme value has no `src` key.
 
 ## [1.2.3] - 2025-11-20